# annotation/documentation/openapi.yaml
openapi: 3.0.2
info:
  title: Badgerdoc Annotation
  version: 0.1.5
paths:
...
  /metadata:
    get:
      tags:
      - Manual Annotation Tasks
      - Metadata
      summary: Get list of possible statuses of tasks.
      operationId: get_entities_statuses_metadata_get
      parameters:
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitiesStatusesSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /jobs/{job_id}:
    post:
      tags:
      - Jobs
      summary: Save info about job.
      operationId: post_job_jobs__job_id__post
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 3
        name: job_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobInfoSchema'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
    patch:
      tags:
      - Jobs
      summary: Update job by job_id.
      operationId: update_job_jobs__job_id__patch
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 1
        name: job_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPatchSchema'
        required: true
      responses:
        '204':
          description: Successful Response
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
...
  /jobs:
    get:
      tags:
      - Jobs
      summary: Get info about jobs, in which provided file ids participate
      operationId: get_jobs_info_by_files_jobs_get
      parameters:
      - required: true
        schema:
          title: File Ids
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
        example:
        - 1
        - 3
        - 4
        name: file_ids
        in: query
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
...
components:
  schemas:
...
    BadRequestErrorSchema:
      title: BadRequestErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Bad request.
...
    ConnectionErrorSchema:
      title: ConnectionErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: 'Error: Connection error.'
...
    EntitiesStatusesSchema:
      title: EntitiesStatusesSchema
      type: object
      properties:
        task_statuses:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusEnumSchema'
          default:
          - Pending
          - Ready
          - In Progress
          - Finished
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    JobInfoSchema:
      title: JobInfoSchema
      required:
      - callback_url
      - annotators
      - validators
      - owners
      - files
      - datasets
      - categories
      - job_type
      type: object
      properties:
        callback_url:
          title: Callback Url
          type: string
          example: http://jobs/jobs/1
        name:
          title: Name
          type: string
          example: job_name
        annotators:
          title: Annotators
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - f0474853-f733-41c0-b897-90b788b822e3
          - b44156f8-e634-48a6-b5f3-c8b1462a2d67
        validators:
          title: Validators
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - f0474853-f733-41c0-b897-90b788b822e3
          - b44156f8-e634-48a6-b5f3-c8b1462a2d67
        owners:
          title: Owners
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - f0474853-f733-41c0-b897-90b788b822e3
          - b44156f8-e634-48a6-b5f3-c8b1462a2d67
        validation_type:
          allOf:
          - $ref: '#/components/schemas/ValidationSchema'
          default: cross
          example: cross
        files:
          title: Files
          uniqueItems: true
          type: array
          items:
            type: integer
          example:
          - 1
          - 2
          - 3
        datasets:
          title: Datasets
          uniqueItems: true
          type: array
          items:
            type: integer
          example:
          - 1
          - 2
          - 3
        is_auto_distribution:
          title: Is Auto Distribution
          type: boolean
          default: false
          example: false
        categories:
          title: Categories
          uniqueItems: true
          type: array
          items:
            type: string
          example:
          - '2'
          - '1'
        deadline:
          title: Deadline
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
        job_type:
          $ref: '#/components/schemas/JobTypeEnumSchema'
...
    JobPatchSchema:
      title: JobPatchSchema
      type: object
      properties:
        callback_url:
          title: Callback Url
          type: string
          example: http://jobs/jobs/1
        name:
          title: Name
          type: string
          example: job_name
        annotators:
          title: Annotators
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - f0474853-f733-41c0-b897-90b788b822e3
        validators:
          title: Validators
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - b44156f8-e634-48a6-b5f3-c8b1462a2d67
        owners:
          title: Owners
          uniqueItems: true
          type: array
          items:
            type: string
            format: uuid
          example:
          - b44156f8-e634-48a6-b5f3-c8b1462a2d67
        files:
          title: Files
          uniqueItems: true
          type: array
          items:
            type: integer
          example:
          - 1
          - 2
          - 3
        datasets:
          title: Datasets
          uniqueItems: true
          type: array
          items:
            type: integer
          example:
          - 1
          - 2
          - 3
        categories:
          title: Categories
          uniqueItems: true
          type: array
          items:
            type: string
          example:
          - '2'
          - '1'
        deadline:
          title: Deadline
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
...
    JobTypeEnumSchema:
      title: JobTypeEnumSchema
      enum:
      - ExtractionJob
      - ExtractionWithAnnotationJob
      - AnnotationJob
      - ImportJob
      type: string
      description: An enumeration.
...
    NotFoundErrorSchema:
      title: NotFoundErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Resource was not found.
...
    TaskStatusEnumSchema:
      title: TaskStatusEnumSchema
      enum:
      - Pending
      - Ready
      - In Progress
      - Finished
      type: string
      description: An enumeration.
...
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    ValidationSchema:
      title: ValidationSchema
      enum:
      - cross
      - hierarchical
      - validation only
      type: string
      description: An enumeration.
...
  securitySchemes:
    TenantDependencyDocs:
      type: http
      scheme: bearer
...